third_party/WebKit/LayoutTests/inspector/console/console-message-contains-async-stack-expected.txt

CONSOLE ERROR: line 1: Uncaught SyntaxError: Unexpected end of input
Tests exception message with empty stack in console contains async stack trace.

VM:1 Uncaught SyntaxError: Unexpected end of input
setTimeout (async)
(anonymous) @ VM:1
